More work on Dew Point.
For implementation of the instrument cavity pressure correction, we need to reverse the processing chain.  This means instead of taking a raw DP and producing a corrected DP and then producing an E from DP (DP->DPC->EDPC), produce E from a raw DP and then a corrected DP from E (DP->E->DPC).  Pressure correction will take place in the computation of E.

In order to support legacy projects and in the interest of consistency across multiple DP measurement instruments, we have also renamed the Buck 1011C variables (example below is for "Left"):

Old	New
DPL	MIRRTMP_DPL
DPLC	DP_DPL
EDPC	EW_DPL

There will now be an E for all probes, and an EWX is introduced for purposes of a reference water vapor pressure.

nimbus will now whether to use Goff-Gratch vs. Murphy-Koop based on the variables names in the default.xml file, DDependTable VarDB will need updates for this if you are backporting to a previous project.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/nimbus@5601 640d5228-2204-0410-8d71-8f46fa6850b4