Fix issue where you can no longer click on timestamp red-dots and get a data buble.
Fix was to reverse the order of the windbarb and timestamp placemarks in the KML file.  This fixes a layering issue where the windbarb hot spot is four times larger than the wind barb thus masking over the little red timestamp placemarks.  Red dots are on top now.

Add timestamp to wind barb bubble.



git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@5667 640d5228-2204-0410-8d71-8f46fa6850b4