Fixed problem when exporting a calibration.
Now prompts the user if they want to insert a duplicate entry when exporting.  The export flag is set regardless of choice.

Cleaned up confusion between database name and table name; both are named 'calibrations'.

Checked in version points to a sandbox version of the calibration table served by ruttles.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@6461 640d5228-2204-0410-8d71-8f46fa6850b4