Checking over this code in the raf/ITAR tree before it is removed from the nidas tree.
Do air alignment when ground speed is over 2 m/s. This was the previous
value in the nidas tree. I had changed it to 5 m/s due to a misreading of
Appendix A (FAQ), section "How do I initialize the unit",
"2. Air Alignment Initialization", where it says no parameters are 
required for initialization when speed is over 5 m/s.

When doing fine stationary alignment, set the validity bit so that the
Auto Align/Nav sequence is used.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@7402 640d5228-2204-0410-8d71-8f46fa6850b4