Merged /branches/ac-track-dp-simplify:r7257-7447
So far this only adds THDG to the AircraftTrack if it exists in the
database, but for that the mapping between query columns and variables
needed refactoring.  Adds tests for comparing kml output when run against
both netcdf files and a database, but the database test data is far from
optimal.  Now requires eol_scons tools.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@7452 640d5228-2204-0410-8d71-8f46fa6850b4