Add support for DMT CIP and PIP probes.
pads2oap has to be run on the DMT PADS files to get them into the generic OAP format.  I beleive it is expecting the ID as C8 for the CIP.  This clearly needs some cleanup to be turnkey.

Work was actually done Spring of 2014 in preperation for OAP processing workshop in Boston.
I want to get this checked in and perhaps it should be in a branch, but Univ Wyoming wants to test it out.



git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@7509 640d5228-2204-0410-8d71-8f46fa6850b4