svn merge -r7601:7779 http://svn.eol.ucar.edu/svn/raf/branches/ac-track-library-python
This merge introduces a new alogrithm to simplify (ie compress the number
of points) KML tracks using the heading instead of time intervals.  The
default is the 'headingstep' algorithm, but it can be changed back to
'timestep' using the --path command-line argument. This merge also includes
automated tests against both netcdf files and real-time SQL databases, and
a way to render web page images of the KML tracks using the same Open
Street Maps code as the mission coordinator web site.  There is (hopefully)
a fix for crossing the anti-meridian by keeping the KML coordinates from
changing sign within the same track.  Finally, as was the original purpose
of this branch, there is a start to a python interface to the aircraft
track library.  See the NOTES.txt file for details.

This closes the ac-track-library-python branch.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@7780 640d5228-2204-0410-8d71-8f46fa6850b4