Use a status handler callback instead of the gotstatus flag, since that's just a better way to do it than toggling a gotstatus flag in the GNISerial object.  If gotstatus did not get reset after being "handled" in GNIServer, then the status could get repeated.  A single flag also limits the possible consumers to one.
Do not pass any fds into the exceptional list in the select() call.  We do
not need to check for any "exceptional conditions" on the fds, since
according to the select_tut(2) man page, we are not interested in any of
them.  Perhaps this will avoid the select() returning immediately and
causing a busy loop because of OOB data or some other exceptional condition
that we don't care about and don't handle.  Just guessing though.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk@8156 640d5228-2204-0410-8d71-8f46fa6850b4