Fix libnimbus so nimbus_tests can link against it.
The global widgets defined in nimbus.c had to move to Xwin.c so they would
be in the library rather than with main(), since otherwise nimbus_tests
would have unresolved symbols when linking with libnimbus.  Turn the
libnimbus settings into a tool, although that turned out not to be strictly
necessary.  This includes some hardcoded logging settings that eventually
need to be removed in favor of command-line options.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/branches/sync-server-trunk-merge@8379 640d5228-2204-0410-8d71-8f46fa6850b4