Add -pr command line option so we can set processing rate on the command line.
This should be an imporvement over setting it in the batch/setup files.  Now we can have one set of batch/setup files and have multiple run_nimbus scripts with the appropriate processing rate set.

-pr rate

Where rate can be 0, 1, 25 or s, l, h; sample-rate, low-rate, or high-rate respectively.

This option will over-ride a "pr=#" in a bath file.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/nimbus@8386 640d5228-2204-0410-8d71-8f46fa6850b4