Code to perform the Atmospheric Temperature retrieval.
At last we have the ability to get physical temperatures given an MTP scan as brightness temperatures, current aircraft altitude and set of tempates to work with.  We also have the first test and output from it.  These values have been compared with the results of the VB MTPBin software and they are identical for the given input.  

This checkin also has:
- improved comments for the basic data structures and methods
- inclusion of all elements of a retrieval coefficient file flight level when the weighted average is calculated.




git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/instruments/mtp@8410 640d5228-2204-0410-8d71-8f46fa6850b4