removed + from gzopen mode
documentation:
gzopen returns NULL if the file could not be opened, if there was insufficient memory to allocate the gzFile state, or if an invalid mode was specified (an 'r', 'w', or 'a' was not provided, or '+' was provided). errno can be checked to determine if the reason gzopen failed was that the file could not be opened.




git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/nimbus@8698 640d5228-2204-0410-8d71-8f46fa6850b4