Add new CVI variables code for icing conditions.
Introduce two new variables, for SOCRATES only at this point.

CVIFLAG to tell when CVI data is good/bad.  This can be expanded in the future to
other critieria for the CVI.

CVCFACTC apply decay factor to CVCFACT if in icing conditions.

When using these, CVCFACTC should be substituted for CVCFACT in the dependencies
for CONCUD and CVCFACTTDL.


git-svn-id: http://svn.eol.ucar.edu/svn/raf/trunk/nimbus@9188 640d5228-2204-0410-8d71-8f46fa6850b4