Fix double free or corruption error on exit.
We have two primary vectors of varaibles RAWTBL and DERTBL.  These are
setup to use a template with a destrutor that will delete the pointers.

We have two other vectors which are subsets of RAWTBL and DERTBL
vectors.  One of these was defined to use template above, causing the
double free.  decode vector was changed to std::vector and not to use
the template above.