Modify computation of UI & VI if ATTACK is nan.
Per Adriana modify horizontal wind computation if ATTACK is nan.

There is a caveat here.  in attack(), ATTACK gets set to AKRD and then anywhere
it is nan it sets the value to 3.0.  This is to help PCORs.  So DependTable for
WI & WIC needs AKRD instead of ATTACK for this to work.