Some additional features of the LoadSetup() facility.
1) We can now have a Flight_All.  This will be applied to all flights, this will be useful for turning off Housekeeping, redundant IRS, ADC, GPS, etc.

2) Adding a PROBE keyword.  So in addition to RAW, DERIVED, DEFAULT, we can now have PROBE.  Probe will be followed by an instrument suffix (e.g. "_RWOO").  This will then loop over all raw and derived variables and perform the desired operation.  Useful for output, data-quality flag, and static lag.  Not so good spike slope, coefficients or dependencies.  This will be best used in the above #1, where you can turn off redundant probes for all flights with a simple "PROBE=_IRS2 O=0"

the LoadSetup() facility works only in Production Mode and will perform the following three actions in the sequence listed:

1) Check the Production/Flight_All file.
2) Check the FlightGroups
3) Check the Production/Flight_rf## <- specific flight number.

Keep in mind it is possible to have a flight setup affected by all three files, it is cumulative, but they will be applied in the order above, so a variable in #3 will override if it appears in #1.  We can have a default Flight_All that is copied in when project is initialized and it will turn off some default redundant probes (_IRS2, _IRS3, _A2, _232, IRIG, A2DTEMP).