Interpolate missing points of the 12.5hz variables.
In ~2016 we had the resampling in phase_shift.c stop interpolating
missing points.  The 12.5hz INS variables had a missing point every
other second to get put into a 13hz sample-rate.  This leads to issues
in high-rate filtering and then in the LATC/LONC/VEWC/VNSC variables at
high-rate.  So for 13hz, remove the missing value that does not actually
exist.