New framwork for metadata
Metadata to be added to the netCDF gets accumulated in a number of
places prior to the netCDF file being created.  Some of this is ad hoc.

 - The common attributes like units & long_names are in the var_base
 - Defaults file attributes are handled by rd_defs.c
 - Values from the PMSsepcs file get added by pms1d.c

We now have a need to add meta_data to other variables (e.g.
std_temperature is read in by an aerosol probe concentration
calculation, but the std_temperure needs to get added to the flow
variable.

I created a Metadata class that holds key/value pairs and is now part of
the var_base class.  NetCDF.cc can now just loop over that as it creates
each variable.

I will refactor the PMSspecs and Defaults into this framework next.

Now all amlib initializers can add there metadata right there.  Prior
there was no place to stash it.