Update pattern matching in Flight setup files.
pattern matching is typically strstr().  But, if it starts with an
underscore, then really it should user strcmp() from the strstr()
position, i.e.  it should match the entire end of the string.

PROBE=_A
was removeing FO3_ACOM
when it should stick to all variables with that exact suffix.