Add support for a SetValues file.
SetValues is similar to BlankOuts, but lets you set a constant
value instead of just nan.Blankouts is a subset of this functionality
most code was reused. This will be useful for data quality flag
variables.

Also corrected spelling of midnight in secondsSinceMidnite

Fixed derived blankouts was not being engaged, removed re-opening of the
file.