Support for the CDP particle by particle (pbp).
The PBPSZ variable has sizes in A2D counts, these need to be converted
to um diameters.

Two ways to go about this.  One is create a derived variable, and the
flow would be good.  Problem, can't get the derived variable out at
sample rate of 2hz, that's only for raw variables.

Two is use an xlate function on the raw variable and overwrite the
sampled data.  This will require putting a return at the top of
xlpbpsz() and recompiling if you want to see the raw A2D values.